adding more files

+++ src/examples/chat/SimpleGlitchSimpleChat.java
@@ -273,21 +273,21 @@ public class SimpleGlitchSimpleChat extends UniversalActor  {
 			{
 				Token token_2_0 = new Token();
 				Token token_2_1 = new Token();
-				// s1<-broadcastSend(new ChatMessage(MessageType.QUESTION, "User one: Poll question one."), 0, 15)
+				// s1<-broadcastSend(new ChatMessage(MessageType.QUESTION, "User one: Poll question one."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.QUESTION, "User one: Poll question one."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.QUESTION, "User one: Poll question one."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s1, "broadcastSend", _arguments, null, token_2_0 );
 					__messages.add( message );
 				}
-				// s1<-broadcastSend(new ChatMessage(MessageType.QUESTION, "User one: Poll question two."), 0, 15)
+				// s1<-broadcastSend(new ChatMessage(MessageType.QUESTION, "User one: Poll question two."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.QUESTION, "User one: Poll question two."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.QUESTION, "User one: Poll question two."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s1, "broadcastSend", _arguments, token_2_0, token_2_1 );
 					__messages.add( message );
 				}
-				// s1<-broadcastSend(new ChatMessage(MessageType.QUESTION, "User one: Poll question three."), 0, 15)
+				// s1<-broadcastSend(new ChatMessage(MessageType.QUESTION, "User one: Poll question three."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.QUESTION, "User one: Poll question three."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.QUESTION, "User one: Poll question three."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s1, "broadcastSend", _arguments, token_2_1, null );
 					__messages.add( message );
 				}
@@ -295,21 +295,21 @@ public class SimpleGlitchSimpleChat extends UniversalActor  {
 			{
 				Token token_2_0 = new Token();
 				Token token_2_1 = new Token();
-				// s2<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User two: Poll response one."), 0, 15)
+				// s2<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User two: Poll response one."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User two: Poll response one."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User two: Poll response one."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s2, "broadcastSend", _arguments, null, token_2_0 );
 					__messages.add( message );
 				}
-				// s2<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User two: Poll response two."), 0, 15)
+				// s2<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User two: Poll response two."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User two: Poll response two."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User two: Poll response two."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s2, "broadcastSend", _arguments, token_2_0, token_2_1 );
 					__messages.add( message );
 				}
-				// s2<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User two: Poll response three."), 0, 15)
+				// s2<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User two: Poll response three."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User two: Poll response three."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User two: Poll response three."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s2, "broadcastSend", _arguments, token_2_1, null );
 					__messages.add( message );
 				}
@@ -317,21 +317,21 @@ public class SimpleGlitchSimpleChat extends UniversalActor  {
 			{
 				Token token_2_0 = new Token();
 				Token token_2_1 = new Token();
-				// s3<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User three: Poll response one."), 0, 15)
+				// s3<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User three: Poll response one."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User three: Poll response one."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User three: Poll response one."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s3, "broadcastSend", _arguments, null, token_2_0 );
 					__messages.add( message );
 				}
-				// s3<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User three: Poll response two."), 0, 15)
+				// s3<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User three: Poll response two."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User three: Poll response two."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User three: Poll response two."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s3, "broadcastSend", _arguments, token_2_0, token_2_1 );
 					__messages.add( message );
 				}
-				// s3<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User three: Poll response three."), 0, 15)
+				// s3<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User three: Poll response three."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User three: Poll response three."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User three: Poll response three."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s3, "broadcastSend", _arguments, token_2_1, null );
 					__messages.add( message );
 				}
@@ -339,21 +339,21 @@ public class SimpleGlitchSimpleChat extends UniversalActor  {
 			{
 				Token token_2_0 = new Token();
 				Token token_2_1 = new Token();
-				// s4<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User four: Poll response one."), 0, 15)
+				// s4<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User four: Poll response one."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User four: Poll response one."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User four: Poll response one."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s4, "broadcastSend", _arguments, null, token_2_0 );
 					__messages.add( message );
 				}
-				// s4<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User four: Poll response two."), 0, 15)
+				// s4<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User four: Poll response two."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User four: Poll response two."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User four: Poll response two."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s4, "broadcastSend", _arguments, token_2_0, token_2_1 );
 					__messages.add( message );
 				}
-				// s4<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User four: Poll response three."), 0, 15)
+				// s4<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User four: Poll response three."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User four: Poll response three."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User four: Poll response three."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s4, "broadcastSend", _arguments, token_2_1, null );
 					__messages.add( message );
 				}
@@ -361,21 +361,21 @@ public class SimpleGlitchSimpleChat extends UniversalActor  {
 			{
 				Token token_2_0 = new Token();
 				Token token_2_1 = new Token();
-				// s5<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User five: Poll response one."), 0, 15)
+				// s5<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User five: Poll response one."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User five: Poll response one."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User five: Poll response one."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s5, "broadcastSend", _arguments, null, token_2_0 );
 					__messages.add( message );
 				}
-				// s5<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User five: Poll response two."), 0, 15)
+				// s5<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User five: Poll response two."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User five: Poll response two."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User five: Poll response two."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s5, "broadcastSend", _arguments, token_2_0, token_2_1 );
 					__messages.add( message );
 				}
-				// s5<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User five: Poll response three."), 0, 15)
+				// s5<-broadcastSend(new ChatMessage(MessageType.ANSWER, "User five: Poll response three."), 3, 12)
 				{
-					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User five: Poll response three."), new Integer(0), new Integer(15) };
+					Object _arguments[] = { new ChatMessage(MessageType.ANSWER, "User five: Poll response three."), new Integer(3), new Integer(12) };
 					Message message = new Message( self, s5, "broadcastSend", _arguments, token_2_1, null );
 					__messages.add( message );
 				}
